+++ src/main/java/online/publicacoes/avaliacaotecnica/repositories/ProcessoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProcessoRepository extends JpaRepository<Processo, Long> {
 
-  Boolean existsByNumeroAndAtivoIsTrue(Integer numero);
+  Boolean existsByNumeroAndAtivoIsTrue(Long numero);
 }
