+++ src/test/java/online/publicacoes/avaliacaotecnica/factories/ProcessoFactory.java
@@ -9,10 +9,10 @@ public class ProcessoFactory {
 
   public static Processo getProcesso() {
 
-    return Processo.builder().numero(1).reu("Réu").ativo(true).build();
+    return Processo.builder().numero(1L).reu("Réu").ativo(true).build();
   }
 
-  public static Processo getProcesso(int n) {
+  public static Processo getProcesso(long n) {
 
     return Processo.builder().numero(n).reu("Réu " + n).ativo(true).build();
   }
